move decoding to separate decoder tile
