+++ crates/relay/src/auctioneer/context.rs
@@ -29,7 +29,7 @@ use crate::{
         simulator::manager::{SimulationResult, SimulatorManager},
         types::{PayloadEntry, PendingPayload, SubmissionRef},
     },
-    spine::messages::SubmissionResultWithRef,
+    spine::{HelixSpineProducers, messages::SubmissionResultWithRef},
 };
 
 // Context that is only valid for a given slot
@@ -126,7 +126,7 @@ impl<B: BidAdjustor> Context<B> {
         &mut self,
         result: SimulationResult,
         already_sent: bool,
-        adapter: &mut flux::spine::SpineAdapter<HelixSpine>,
+        producers: &mut HelixSpineProducers,
     ) {
         let (id, result) = result;
 
@@ -188,7 +188,7 @@ impl<B: BidAdjustor> Context<B> {
 
         if !already_sent && !result.optimistic_version.is_optimistic() {
             send_submission_result(
-                &mut adapter.producers,
+                producers,
                 &self.future_results,
                 result.submission_ref,
                 Err(BuilderApiError::SimOnNextSlot),
