+++ crates/relay/src/main.rs
@@ -25,11 +25,7 @@ use helix_common::{
     utils::{init_panic_hook, init_tracing_log},
 };
 use helix_relay::{
-    Api, Auctioneer, AuctioneerHandle, BidSorter, BidSubmissionTcpListener, DbHandle,
-    DefaultBidAdjustor, FutureBidSubmissionResult, HelixSpine, InternalBidSubmission, RegWorker,
-    RegWorkerHandle, RelayNetworkManager, S3PayloadSaver, SubWorker, SubmissionResultWithRef,
-    SubmissionResultsFanOut, WebsiteService, spawn_tokio_monitoring, start_admin_service,
-    start_api_service, start_beacon_client, start_db_service, start_housekeeper,
+    Api, Auctioneer, AuctioneerHandle, BidSorter, BidSubmissionTcpListener, DbHandle, DecoderTile, DefaultBidAdjustor, FutureBidSubmissionResult, HelixSpine, InternalBidSubmission, RegWorker, RegWorkerHandle, RelayNetworkManager, S3PayloadSaver, SubWorker, SubmissionDataWithSpan, SubmissionResultWithRef, SubmissionResultsFanOut, WebsiteService, spawn_tokio_monitoring, start_admin_service, start_api_service, start_beacon_client, start_db_service, start_housekeeper
 };
 use helix_types::BlsKeypair;
 use tikv_jemallocator::Jemalloc;
@@ -157,6 +153,9 @@ async fn run(instance_id: String, config: RelayConfig, keypair: BlsKeypair) -> e
         let submission_results = Arc::new(SharedVector::<SubmissionResultWithRef>::with_capacity(
             MAX_SUBMISSIONS_PER_SLOT,
         ));
+        let decoded = Arc::new(SharedVector::<SubmissionDataWithSpan>::with_capacity(
+            MAX_SUBMISSIONS_PER_SLOT,
+        ));
 
         let bid_producer = spine.spine.standalone_producer_for(TileName::from_str_truncate("Api"));
         start_api_service::<ApiProd>(
@@ -194,16 +193,23 @@ async fn run(instance_id: String, config: RelayConfig, keypair: BlsKeypair) -> e
         }
 
         if config.is_submission_instance {
+            // TODO multiple decoder tiles
+            let decoder_tile = DecoderTile::new(
+                local_cache.as_ref().clone(),
+                chain_info.as_ref().clone(), 
+                config.clone(), 
+                submissions.clone(), 
+                submission_results.clone(), 
+                decoded.clone(),
+            );
+            attach_tile(decoder_tile, spine, TileConfig::new(config.cores.decoder, ThreadPriority::OSDefault));
+
             for core in config.cores.sub_workers.clone() {
                 let worker = SubWorker::new(
                     core,
                     event_tx.clone(),
                     sub_worker_rx.clone(),
-                    local_cache.as_ref().clone(),
                     chain_info.as_ref().clone(),
-                    config.clone(),
-                    submissions.clone(),
-                    future_results.clone(),
                 );
 
                 attach_tile(worker, spine, TileConfig::new(core, ThreadPriority::OSDefault));
@@ -255,6 +261,7 @@ async fn run(instance_id: String, config: RelayConfig, keypair: BlsKeypair) -> e
                 event_rx,
                 auctioneer_core,
                 future_results,
+                decoded,
             );
             attach_tile(
                 auctioneer,
